# openapi.yaml
openapi: 3.0.3
info:
  title: HabitatTrack Financial Management API
  version: v1.0.0
  description: API for managing properties, financial transactions, and categories.
  contact:
    name: API Support
    url: http://www.example.com/support
    email: support@example.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.habitattrack.com
    description: Production server
components:
  schemas:
...
    TransactionCategory:
      type: object
      required:
        - name
        - classification
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the transaction category.
          readOnly: true
        name:
          type: string
          description: Name of the transaction category (e.g., "Rent Received", "Insurance Premium").
          example: "Utility Bill"
        classification:
          type: string
          enum: [income, expense]
          description: Classification of the category.
          example: "expense"
        created_at:
          type: string
          format: date-time
          description: Timestamp of when the category was created.
          readOnly: true
        updated_at:
          type: string
          format: date-time
          description: Timestamp of when the category was last updated.
          readOnly: true
    TransactionCategoryCreateRequest:
      type: object
      required:
        - name
        - classification
      properties:
        name:
          type: string
          description: Name of the transaction category.
        classification:
          type: string
          enum: [income, expense]
          description: Classification of the category.
    TransactionCategoryUpdateRequest:
      type: object
      properties:
        name:
          type: string
          description: Name of the transaction category.
        classification:
          type: string
          enum: [income, expense]
          description: Classification of the category.
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: An application-specific error code.
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: A human-readable message providing more details about the error.
          example: "The 'name' field is required."
        details:
          type: object
          additionalProperties:
            type: string
          description: Optional field for providing more specific error details, e.g., field-level validation errors.
          example:
            field: "name"
            issue: "cannot be empty"
    PaginationInfo:
      type: object
      properties:
        totalItems:
          type: integer
          format: int64
          description: Total number of items available.
        totalPages:
          type: integer
          format: int32
          description: Total number of pages.
        currentPage:
          type: integer
          format: int32
          description: The current page number.
        pageSize:
          type: integer
          format: int32
          description: Number of items per page.
        nextCursor:
          type: string
          nullable: true
          description: Cursor for fetching the next page (for cursor-based pagination).
        prevCursor:
          type: string
          nullable: true
          description: Cursor for fetching the previous page (for cursor-based pagination).
...
    PaginatedCategoriesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TransactionCategory'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
  parameters:
...
    CategoryID:
      name: categoryId
      in: path
      required: true
      description: Identifier of the transaction category.
      schema:
        type: string
        format: uuid
...
    PageLimit:
      name: limit
      in: query
      description: Maximum number of items to return per page.
      required: false
      schema:
        type: integer
        format: int32
        default: 20
        minimum: 1
        maximum: 100
    PageOffset:
      name: offset
      in: query
      description: Number of items to skip before starting to collect the result set.
      required: false
      schema:
        type: integer
        format: int32
        default: 0
        minimum: 0
    # Pagination Parameters (Cursor-based example)
...
  securitySchemes:
    OAuth2BearerAuth:
      type: oauth2
      description: OAuth 2.0 Bearer Token for API authentication.
      flows:
        implicit: # Or clientCredentials, authorizationCode, password depending on your auth strategy
          authorizationUrl: https://auth.habitattrack.com/oauth2/authorize
          scopes:
            read:transactions: Grants read access to transactions
            write:transactions: Grants write access to transactions
            read:categories: Grants read access to categories
            write:categories: Grants write access to categories
            read:properties: Grants read access to properties
            write:properties: Grants write access to properties
# Global security requirement (can be overridden at operation level)
security:
  - OAuth2BearerAuth: [] # Empty array means no specific scopes required globally, define per operation if needed
paths:
...
  /transaction-categories:
    summary: Manage transaction categories.
    get:
      summary: List all transaction categories
      tags: [Categories]
      operationId: listTransactionCategories
      parameters:
        - name: classification
          in: query
          description: Filter categories by classification (income or expense).
          required: false
          schema:
            type: string
            enum: [income, expense]
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageOffset'
      responses:
        '200':
          description: A paginated list of transaction categories.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCategoriesResponse'
        '400':
          description: Invalid filter parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - OAuth2BearerAuth: [read:categories]
    post:
      summary: Create a new transaction category
      tags: [Categories]
      operationId: createTransactionCategory
      requestBody:
        description: Transaction category object to be created.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryCreateRequest'
      responses:
        '201':
          description: Transaction category created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionCategory'
          headers:
            Location:
              description: URL of the newly created category.
              schema:
                type: string
                format: url
        '400':
          description: Invalid input (e.g., validation error).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict (e.g., category name already exists for the given classification).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - OAuth2BearerAuth: [write:categories]
  /transaction-categories/{categoryId}:
    summary: Manage a specific transaction category.
    parameters:
      - $ref: '#/components/parameters/CategoryID'
    get:
      summary: Get a transaction category by its ID
      tags: [Categories]
      operationId: getTransactionCategoryById
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionCategory'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transaction category not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - OAuth2BearerAuth: [read:categories]
    put:
      summary: Update an existing transaction category (replaces the entire resource)
      tags: [Categories]
      operationId: updateTransactionCategory
      requestBody:
        description: Transaction category object to update.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryCreateRequest' # All fields required for replacement
      responses:
        '200':
          description: Transaction category updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionCategory'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transaction category not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict (e.g., category name already exists for the given classification).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - OAuth2BearerAuth: [write:categories]
    patch:
      summary: Partially update an existing transaction category
      tags: [Categories]
      operationId: patchTransactionCategory
      requestBody:
        description: Transaction category fields to update.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryUpdateRequest'
      responses:
        '200':
          description: Transaction category partially updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionCategory'
        '400':
          description: Invalid input for patch operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transaction category not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict (e.g., category name already exists for the given classification).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - OAuth2BearerAuth: [write:categories]
    delete:
      summary: Delete a transaction category by its ID
      tags: [Categories]
      operationId: deleteTransactionCategory
      responses:
        '204':
          description: Transaction category deleted successfully. No content.
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Transaction category not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict (e.g., category is in use by transactions and cannot be deleted).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - OAuth2BearerAuth: [write:categories]
...